refactor: remove dependency from restapi dir.
